+++ example/lib/screens/stack_variant/single_choice/s_stack_screen.dart
@@ -27,7 +27,7 @@ class _SStackScreenState extends State<SStackScreen> {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      appBar: AppBar(title: const Text("StackWidget (multiple choice)")),
+      appBar: AppBar(title: const Text("StackWidget (single choice)")),
       body: SafeArea(
         child: Column(
           children: [