+++ lib/views/expanded_variant/multiple_choice/m_lazy_expanded_widget.p.dart
@@ -3,33 +3,36 @@
  * Email: nvbien2000@gmail.com
  */
 
-part of '../../recursive_tree_flutter.dart';
+part of '../../../recursive_tree_flutter.dart';
 
-class ExpandedWidget<T extends AbsNodeType> extends StatefulWidget {
-  const ExpandedWidget(
+class MLazyExpandedWidget<T extends AbsNodeType> extends StatefulWidget {
+  const MLazyExpandedWidget(
     this.initData, {
     super.key,
     this.properties = const UIProperties(),
+    required this.fGetChildrenFunc,
   });
 
   final TreeType<T> initData;
   final UIProperties properties;
+  final FGetChildrenFunc<T> fGetChildrenFunc;
 
   @override
-  State<ExpandedWidget<T>> createState() => _ExpandedWidgetState<T>();
+  State<MLazyExpandedWidget<T>> createState() => _MLazyExpandedWidgetState<T>();
 }
 
-class _ExpandedWidgetState<T extends AbsNodeType>
-    extends State<ExpandedWidget<T>> {
+class _MLazyExpandedWidgetState<T extends AbsNodeType>
+    extends State<MLazyExpandedWidget<T>> {
   @override
   Widget build(BuildContext context) {
     return ListView.builder(
       itemCount: 1,
       itemBuilder: (_, __) {
-        return _EIWidget<T>(
+        return _MLazyEIWidget<T>(
           widget.initData,
           setStateEntireWidget: () => setState(() {}),
           properties: widget.properties,
+          fGetChildrenFunc: widget.fGetChildrenFunc,
         );
       },
     );